code consolidation, adds ability to reselect location
