+++ app/src/main/java/com/smashingboxes/epa_prototype_android/AppStateManager.java
@@ -4,6 +4,10 @@ import android.content.Context;
 
 import com.smashingboxes.epa_prototype_android.helpers.PreferenceHelper;
 import com.smashingboxes.epa_prototype_android.fitbit.location.SimplePlace;
+import com.smashingboxes.epa_prototype_android.network.epa.models.EpaActivity;
+
+import java.util.HashMap;
+import java.util.Map;
 
 /**
  * Created by Austin Lanier on 12/16/15.
@@ -27,6 +31,7 @@ public class AppStateManager {
 
     private PreferenceHelper mPreferenceHelper;
     private SimplePlace place;
+    private Map<String, EpaActivity.Location> mLocationMap = new HashMap<>();
 
     public SimplePlace getPlace() {
         if (place == null) {
@@ -37,11 +42,30 @@ public class AppStateManager {
 
     public void savePlace(SimplePlace place) {
         this.place = place;
-        mPreferenceHelper.persistObjectSync(PLACE_KEY, place);
+        mPreferenceHelper.persistObjectAsync(PLACE_KEY, place);
     }
 
     public void clearPlace() {
         mPreferenceHelper.removeObjectForKey(PLACE_KEY);
+    }
+
+    public void addLocation(String date, EpaActivity.Location location) {
+        mPreferenceHelper.persistStringAsync(date, location.name());
+        mLocationMap.put(date, location);
+    }
+
+    public EpaActivity.Location getLocation(String date) {
+        EpaActivity.Location location = mLocationMap.get(date);
+        if (location == null) {
+            location = EpaActivity.Location.valueOf(mPreferenceHelper.getString(date, EpaActivity.Location.NONE.name()));
+        }
+        return location;
+    }
+
+    public void clear() {
+        clearPlace();
+        mLocationMap.clear();
+        mPreferenceHelper.clear();
     }
 
 }
