+++ app/src/main/res/values/ids.xml
@@ -7,6 +7,7 @@
 	<item name="web_view" type="id"/>
 	<item name="menu_logout" type="id"/>
 	<item name="menu_settings" type="id"/>
+	<item name="menu_select_location" type="id"/>
 	<item name="user_image" type="id"/>
 	<item name="user_details_container" type="id"/>
 	<item name="sign_in_button" type="id"/>
@@ -22,5 +23,6 @@
 	<item name="swipe_container" type="id"/>
 	<item name="activity_inside" type="id"/>
 	<item name="activity_outside" type="id"/>
+
 
 </resources>
